added intersection of polygons
