+++ src/polygon/polygon.cpp
